CA-59: add rubocop
