--- app/views/events/_event.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! event, :id, :name, :time_id, :recommend_menu, :place, :max_num, :created_at, :updated_at
-json.url event_url(event, format: :json)
