--- app/views/events/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @events, partial: "events/event", as: :event
