--- app/views/user_mailer/sample.html.erb
@@ -1,5 +1,0 @@
-@user.name 様
-
-    会員登録が完了しました。
-    URLはこちらです。
-<%= user_url(@user) %>
