+++ spec/factories/time_tables.rb
@@ -1,6 +1,5 @@
 FactoryBot.define do
     factory :time_table do
-        id {1}
         time {"18:00 ~ 20:00"}
     end
 end
